# Optimize String Split