--- a/benchmark/micro/list/string_split_regexp.benchmark
+++ b/benchmark/micro/list/string_split_regexp.benchmark
@@ -1,0 +1,20 @@
+# name: benchmark/micro/list/string_split_regexp.benchmark
+# description: String split regexp benchmark
+# group: [list]
+
+name String Split Regexp
+group micro
+subgroup list
+
+require tpch
+
+cache tpch_sf1
+
+load
+CALL dbgen(sf=1);
+
+run
+SELECT SUM(LENGTH(str_split_regex(l_comment, '[z ]'))) FROM lineitem
+
+result I
+27186473