--- a/src/function/scalar/string/string_split.cpp
+++ b/src/function/scalar/string/string_split.cpp
@@ -81,7 +81,7 @@ struct StringSplitter {
 			if (match_size == 0 && pos == 0) {
 				// special case: 0 length match and pos is 0
 				// move to the next character
-				for(pos++; pos < input_size; pos++) {
+				for (pos++; pos < input_size; pos++) {
 					if (LengthFun::IsCharacter(input_data[pos])) {
 						break;
 					}