--- a/test/sql/function/string/test_string_split.test
+++ b/test/sql/function/string/test_string_split.test
@@ -96,15 +96,23 @@ SELECT UNNEST(string_split('', 'delim'))
 ----
 (empty)
 
-statement error
+query T
 SELECT UNNEST(string_split('aaaaa', ''))
 ----
-empty separator
+a
+a
+a
+a
+a
 
-statement error
+query T
 SELECT UNNEST(string_split('🦆🦆🦆🦆🦆', ''))
 ----
-empty separator
+🦆
+🦆
+🦆
+🦆
+🦆
 
 query T
 SELECT UNNEST(string_split('abab', 'b'))
@@ -166,10 +174,14 @@ a
 a
 a
 
-statement error
+query T
 SELECT UNNEST(string_split_regex('aaaaa', ''))
 ----
-empty separator
+a
+a
+a
+a
+a
 
 
 query T
@@ -208,10 +220,14 @@ select UNNEST(string_split('1||2|3||', '||'))
 2|3
 (empty)
 
-statement error
+query T
 select UNNEST(string_split('1|2|3', ''))
 ----
-empty separator
+1
+|
+2
+|
+3
 
 query T
 select UNNEST(string_split('', '|'))
@@ -228,19 +244,19 @@ select string_split(NULL, '|') IS NULL
 ----
 1
 
-statement error
+query T
 select UNNEST(string_split('abc', ''))
 ----
-empty separator
+a
+b
+c
 
 query T
 select UNNEST(string_split_regex('abc', '(|abc)'))
 ----
-(empty)
 a
 b
 c
-(empty)
 
 query T
 select UNNEST(string_split_regex('abc', '(abc|)'))
@@ -266,41 +282,6 @@ select UNNEST(string_split_regex('abc', '(abc|,)'))
 (empty)
 
 query T
-select UNNEST(string_split_regex('axxxc', 'z*'))
-----
-(empty)
-a
-x
-x
-x
-c
-(empty)
-
-query T
-select UNNEST(string_split(' aa', ' '))
-----
-(empty)
-aa
-
-query T
-select UNNEST(string_split(' aa ', ' '))
-----
-(empty)
-aa
-(empty)
-
-query T
-select UNNEST(string_split('aa ', ' '))
-----
-aa
-(empty)
-
-query T
-select UNNEST(string_split_regex('axxxc', 'z+'))
-----
-axxxc
-
-query T
 select UNNEST(string_split('1,2,3,4,,6', ','))
 ----
 1
@@ -313,7 +294,6 @@ select UNNEST(string_split('1,2,3,4,,6', ','))
 query T
 select UNNEST(string_split_regex('1,2,3,4,,6', '(,|)'))
 ----
-(empty)
 1
 (empty)
 2
@@ -324,13 +304,11 @@ select UNNEST(string_split_regex('1,2,3,4,,6', '(,|)'))
 (empty)
 (empty)
 6
-(empty)
 
 
 query T
 select UNNEST(string_split_regex('1,2,3,4,,6', '(|,)'))
 ----
-(empty)
 1
 ,
 2
@@ -341,7 +319,6 @@ select UNNEST(string_split_regex('1,2,3,4,,6', '(|,)'))
 ,
 ,
 6
-(empty)
 
 query T
 select UNNEST(string_split_regex('1,2,3,4,*,6', '(,|\*)'))
@@ -364,14 +341,6 @@ select UNNEST(string_split_regex('1,2,3,4,*,6', '(\*|,)'))
 (empty)
 (empty)
 6
-
-query T
-SELECT UNNEST(string_split_regex('-AA=AA+aA*', 'a+', 'i'))
-----
--
-=
-+
-*
 
 # test incorrect usage
 statement error