This PR reworks the string split operator and optimizes it significantly, particularly the `regexp_string_split` operator sees large performance improvements. In summary:

* Avoid using the Value API as much as possible, and directly construct results as `string_t` objects
* Simplify implementation of string_split to use existing optimized search used in `Contains`, rather than re-implementing the string search in a slower manner
* For regexp_string_split, cache the regexp if it is a constant to avoid having to re-compile it for every value (which is the most common case)

Some benchmarks:

###### ASCII string split

```sql
CALL dbgen(sf=1);
CREATE TABLE duck_comments AS SELECT concat(l_comment, '🦆') l_comment FROM lineitem;
SELECT SUM(LENGTH(str_split(l_comment, ' '))) FROM lineitem;
```

###### Unicode string split
```sql
CALL dbgen(sf=1);
SELECT SUM(LENGTH(str_split(l_comment, ' '))) FROM duck_comments;
```

###### Regex string split
```sql
CALL dbgen(sf=1);
SELECT SUM(LENGTH(str_split_regex(l_comment, '[z ]'))) FROM lineitem
```

| Benchmark |  Old  |  New  |
|-----------|-------|-------|
| ASCII     | 0.72s | 0.08s |
| Unicode   | 0.91s | 0.08s |
| Regex     | 17.5s | 0.29s |