@Mytherin 2022-11-03
Hah, yes, I figured :) Back then the nested types were also a lot less well defined and understood.